Extract syscall_read() handler
